refactor(db): database/init.ts 스키마 초기화 모듈 분해 (#631)
713줄 단일 파일을 책임별 sub-module로 분리해 오케스트레이터를 500줄 이하로 유지하고, public import 경로와 동작은 그대로 둔다.
